alus-147 SRTM3 all tile indexes support.
* Refactored existing tests to always include SRTM3 tile count argument
* Refactored existing code to supply SRTM3 tile count argument
* Creates ID based on SRTM3 DEM file location on earth which is used
inside the device code. For example a SRTM3 tile with a name
srtm_30_07.tif would result in an ID of 3007. This ID is used to
select correct DEM value array inside kernel code. Since originally
this is quite direct copy of SNAP therefore it could be refactored
so that no ID based segregation is needed to be done in future.